renamed slidehshow folder
